Require pyinstaller after all
I wasn't putting it in before because it's not really needed unless
you're creating standalone apps, but upon further reflection, having
a standalone app is kind of a key feature of this package, and I don't
expect anyone to install it as a python module per se.  So we should
include pyinstaller as a requirement in that sense.