Backed out informing user they can quit
The way it was implemented meant it would show up when using the CLI
too.  This was confusing and pointless in that situation.  The problem
is, there's no way for either RunManager or UI itself to put up the
message at the right time; the only way is in MainBody because it
knows when it's done and the time is right.  But MainBody doesn't know
whether a CLI or GUI is being used.

I don't have a good solution to this dilemma so I'm leaving it for now
-- it's not important enough to spend more time on it.